Address comments from Will H
