+++ packages/node/src/constants.ts
@@ -1,4 +1,3 @@
 export const SDK_NAME = 'amplitude-node';
 export const SDK_VERSION = '0.3.2';
-export const AMPLITUDE_API_HOST = 'api.amplitude.com';
-export const AMPLITUDE_API_PATH = '/2/httpapi';
+export const AMPLITUDE_SERVER_URL = 'https://api2.amplitude.com/2/httpapi';
