+++ packages/node/src/transports/index.ts
@@ -1,3 +1,1 @@
-export { BaseTransport } from './base';
 export { HTTPTransport } from './http';
-export { HTTPSTransport } from './https';
