+++ packages/types/src/index.ts
@@ -1,7 +1,6 @@
 export { Client } from './client';
-export { Event } from './event';
+export { Event, Payload } from './event';
 export { Options } from './options';
-export { Payload } from './payload';
 export { Response } from './response';
 export { Status } from './status';
 export { Transport, TransportOptions } from './transport';
