--- packages/types/src/payload.ts
@@ -1,9 +1,0 @@
-import { Event } from './event';
-
-/**
- * Amplitude request payload definition.
- */
-export interface Payload {
-  api_key: string;
-  events: Event[];
-}
