+++ packages/types/src/response.ts
@@ -3,4 +3,5 @@ import { Status } from './status';
 /** JSDoc */
 export interface Response {
   status: Status;
+  statusCode: number;
 }
